Move useTemplatesApi hook into separate file
